feat: optimize bit mask operations

+++ src/insts/format.rs
@@ -31,7 +31,7 @@ pub fn b(f: impl Fn(usize, usize, i64) -> Inst, inst32: u32) -> Inst {
 
 #[inline]
 pub fn u(f: impl Fn(usize, i64) -> Inst, inst32: u32) -> Inst {
-    let imm_31_12 = ext!(inst32, u128; 31;12);
+    let imm_31_12 = ext!(inst32, u64; 31;12);
     let imm = sext(imm_31_12 << 12, 31) as i64;
     f(rd_11_7(inst32), imm)
 }
